+++ packages/compiler/examples/react-app/src/index.jsx
@@ -1,11 +1,12 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App';
-import registerServiceWorker from './registerServiceWorker';
 
-ReactDOM.render(<App />, document.getElementById('root'));
-registerServiceWorker();
+const container = document.getElementById('root');
+
+const root = createRoot(container);
+
+root.render(<App />);
 
 if (import.meta.webpackHot) {
   import.meta.webpackHot.accept();