+++ packages/compiler/examples/ts-css-modules/src/styles.module.scss.d.ts
@@ -1,12 +1,5 @@
-declare namespace StylesModuleScssNamespace {
-  export interface IStylesModuleScss {
-    block: string;
-  }
+export interface IStylesModuleScss {
+  'block': string;
 }
-
-declare const StylesModuleScssModule: StylesModuleScssNamespace.IStylesModuleScss & {
-  /** WARNING: Only available when `css-loader` is used without `style-loader` or `mini-css-extract-plugin` */
-  locals: StylesModuleScssNamespace.IStylesModuleScss;
-};
-
-export = StylesModuleScssModule;
+declare const styles: IStylesModuleScss;
+export = styles;