+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ## [8.0.0]
 
 - Updated all dependencies
+- Fixed d.ts generation for CSS Modules
+- Improved debugging in production
 - Dropped handlebars and pug support
 - Dropped outdated markupCompiler
+- Dropped outdated webpack plugins
 - Dropped bower support
 
 ## [7.2.0]